Feat #364/front custom filters (#676)
* feat: :sparkles: add custom filters

* wip

* Custom filters are working. Adding, updating, deleting and renaming custom filters

* fix: CreateCustomFilterDto omit id

* fix: refresh ag-grid

* cleaning in case its about columns that doesnt exist

* now managing column selection with aggr id

* managing column order

* correcting test e2e back

---------

+++ apps/server/src/modules/custom-filters/controllers/custom-filter.controller.ts
@@ -1,4 +1,15 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post, UseGuards, UseInterceptors } from '@nestjs/common'
+import {
+  BadRequestException,
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Patch,
+  Post,
+  UseGuards,
+  UseInterceptors,
+} from '@nestjs/common'
 import { ApiTags } from '@nestjs/swagger'
 import { LoggerService } from '@/shared/modules/logger/logger.service'
 import { CustomFilter } from '@/modules/custom-filters/objects/entities/custom-filter.entity'
@@ -60,6 +71,9 @@ export class CustomFilterController {
   async deleteOneFilter(@Param('filterId') filterId: number,
                         @CurrentUserId() userId: number): Promise<DeleteResult> {
     this.logger.verbose('deleteOneFilter')
+    if (!Number(filterId)) {
+      throw new BadRequestException('Filter id must be a number')
+    }
     return this.service.remove({ id: filterId, userId })
   }
 }
